CXP-1291: moved useChannel  to the shared namespace
