+++ components/catalogs/front/src/components/ProductSelection/criteria/CompletenessCriterion/CompletenessScopeInput.tsx
@@ -2,7 +2,7 @@ import React, {FC} from 'react';
 import {SelectInput} from 'akeneo-design-system';
 import {CompletenessCriterionState} from './types';
 import {useTranslate} from '@akeneo-pim-community/shared';
-import {useChannel} from '../../hooks/useChannel';
+import {useChannel} from '../../../../hooks/useChannel';
 import {Channel} from '../../../../models/Channel';
 import {useInfiniteChannels} from '../../../../hooks/useInfiniteChannels';
 import {useUniqueEntitiesByCode} from '../../../../hooks/useUniqueEntitiesByCode';
